Create README.md
In this test task, I demonstrate my skills as a Middle Developer by implementing the MVVM pattern. I am experimenting with fragments to create a faster and more responsive interface compared to using Activities. Additionally, this assignment showcases my proficiency in working with background services and the Google Maps API.